- added a --leaderboard option to Grabby mode. it allows to archive multiple places based on leaderboards. requires specifying leaderboard category and period (currently only one of each is supported - see below). alters default --out2 saving path and appends a new --out3 path available only in this mode. - added a leaderboard category 'Regions'. enabled by setting --by-region. only available in Grabby mode with --leaderboard turned on. all the regions are hardcoded due to lack of API endpoint on wplace to fetch them from. - added a leaderboard period 'all-time'. enabled by setting --all-time. only available in Grabby mode with --leaderboard turned on. - grabby tile position is now optional when in --leaderboard mode. - --out/--out2 variables are now specified with '%' prefix. - added a --cycle-start-delay option to control delay before a cycle starts. - improved descriptions for --out/--out2 options - enabled minify and sourcemap options for builds. builds should be smaller now and with proper stacktraces in case of errors. - made timeout error retryable in tile fetch queue in case abort timeout ever increases and goes past the normal timeout duration. - moved process.exit() out of grabby up to cli handler so it doesn't quit mid leaderboard save. the exit() is needed for grabby because there exists a mysterious dangling promise that doesn't allow natural exiting nor it throws any errors when forcebly throwing errors...

+++ src/lib/Cycler.ts
@@ -99,12 +99,15 @@ export class Cycler {
             errorsDirpathRelToWorkDir = `${outDirpathRelToWorkDir}-ERRORS`;
             errorsDirpath = path.join(this._workingDir, errorsDirpathRelToWorkDir);
 
-            fs.renameSync(oldOutDirpath, outDirpath);
+            if (outDirpath !== oldOutDirpath)
+                fs.renameSync(oldOutDirpath, outDirpath);
 
-            if ((await fs.readdir(oldErrorsDirpath)).length === 0)
-                fs.rmdirSync(oldErrorsDirpath);
-            else
-                fs.renameSync(oldErrorsDirpath, errorsDirpath);
+            if (errorsDirpath !== oldErrorsDirpath) {
+                if ((await fs.readdir(oldErrorsDirpath)).length === 0)
+                    fs.rmdirSync(oldErrorsDirpath);
+                else
+                    fs.renameSync(oldErrorsDirpath, errorsDirpath);
+            }
 
             const elapsedFmted = humanizeDuration(elapsedMs, { round: true });
             if (loop) {
